Move perf functions in a shared utils file
